+++ src/Direction/CollectionDirection.php
@@ -5,14 +5,14 @@ declare(strict_types=1);
 namespace Yansongda\Pay\Direction;
 
 use Psr\Http\Message\ResponseInterface;
+use Yansongda\Pay\Contract\DirectionInterface;
 use Yansongda\Pay\Contract\PackerInterface;
-use Yansongda\Pay\Contract\ParserInterface;
 use Yansongda\Pay\Exception\ContainerException;
 use Yansongda\Pay\Exception\ServiceNotFoundException;
 use Yansongda\Pay\Pay;
 use Yansongda\Supports\Collection;
 
-class CollectionDirection implements ParserInterface
+class CollectionDirection implements DirectionInterface
 {
     /**
      * @throws ContainerException