+++ src/Direction/NoHttpRequestDirection.php
@@ -5,10 +5,10 @@ declare(strict_types=1);
 namespace Yansongda\Pay\Direction;
 
 use Psr\Http\Message\ResponseInterface;
+use Yansongda\Pay\Contract\DirectionInterface;
 use Yansongda\Pay\Contract\PackerInterface;
-use Yansongda\Pay\Contract\ParserInterface;
 
-class NoHttpRequestDirection implements ParserInterface
+class NoHttpRequestDirection implements DirectionInterface
 {
     public function parse(PackerInterface $packer, ?ResponseInterface $response): ?ResponseInterface
     {