+++ src/Direction/OriginResponseDirection.php
@@ -5,12 +5,12 @@ declare(strict_types=1);
 namespace Yansongda\Pay\Direction;
 
 use Psr\Http\Message\ResponseInterface;
+use Yansongda\Pay\Contract\DirectionInterface;
 use Yansongda\Pay\Contract\PackerInterface;
-use Yansongda\Pay\Contract\ParserInterface;
 use Yansongda\Pay\Exception\Exception;
 use Yansongda\Pay\Exception\InvalidResponseException;
 
-class OriginResponseDirection implements ParserInterface
+class OriginResponseDirection implements DirectionInterface
 {
     /**
      * @throws InvalidResponseException