+++ CHANGELOG.md
@@ -25,7 +25,7 @@
 - change: 插件开始装载日志由 `info` 调整为 `debug`(#755)
 - change: ParserInterface 签名由 `?ResponseInterface $response` 变更为 `PackerInterface $packer, ?ResponseInterface $response`(#754)
 - change: \Yansongda\Pay\Plugin\Wechat\RadarSignPlugin 增加 `__construct(JsonPacker $jsonPacker, XmlPacker $xmlPacker)` 方法(#753)
-- change: 所有 `Parser` 更名为 `Direction`(#770)
+- change: 所有 `Parser` 更名为 `Direction`(#770, #774)
 
 ## v3.2.14